[#108] Refactored add vocabulary form to multiple files
